requirements.txt
# 画像さがす君（ローカルAI画像検索システム）依存ライブラリ
# Version 2.0

# Google Gemini API
google-generativeai>=0.3.0

# 機械学習・深層学習
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# 画像処理
Pillow>=9.0.0
opencv-python>=4.8.0

# 数値計算・科学技術計算
numpy>=1.21.0
scipy>=1.10.0
scikit-learn>=1.0.0

# 日本語処理
janome>=0.4.0

# ONNX Runtime（NPU/GPU対応）
onnxruntime>=1.15.0

# その他のユーティリティ
requests>=2.28.0
tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

# 開発用（オプション）
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0 